210210 노마드코더 firebase 트위터 클론코딩 마무리

+++ src/components/Nweet.js
@@ -10,7 +10,7 @@ const Nweet = ({ nweetObj, isOwner }) => {
     const ok = window.confirm("Are you sure you want delete this nweet?")
     if(ok) {
       await dbService.doc(`nweets/${nweetObj.id}`).delete()
-      await storageService.refFromURL(nweetObj.attachmentUrl).delete()
+      if(nweetObj.attachmentUrl !== "") await storageService.refFromURL(nweetObj.attachmentUrl).delete()
     }
   }
   const toggleEditing = () => setEditing(prev => !prev)
